+++ examples/tool-calling/CHANGELOG.md
@@ -5,6 +5,6 @@
 ### Patch Changes
 
 - Updated dependencies [f4c92e5]
-  - @strand/anthropic@0.1.0
-  - @strand/core@0.1.0
-  - @strand/react@0.1.0
+  - @strandjs/anthropic@0.1.0
+  - @strandjs/core@0.1.0
+  - @strandjs/react@0.1.0